+++ src/Console/PaperCommand.php
@@ -9,6 +9,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Filesystem\Filesystem;
 use JacobJoergensen\LaravelPaper\Contracts\PaperModel;
 use JacobJoergensen\LaravelPaper\Exceptions\PaperException;
+use JacobJoergensen\LaravelPaper\Paper;
 use ReflectionClass;
 
 /**
@@ -136,6 +137,6 @@ abstract class PaperCommand extends Command
      */
     private function isPaperModel(string $model): bool
     {
-        return is_subclass_of($model, PaperModel::class);
+        return class_exists($model) && isset(class_uses_recursive($model)[Paper::class]);
     }
 }