+++ tests/Fixtures/Article.php
@@ -4,20 +4,16 @@ declare(strict_types=1);
 
 namespace JacobJoergensen\LaravelPaper\Tests\Fixtures;
 
-use Illuminate\Database\Eloquent\Model;
 use JacobJoergensen\LaravelPaper\Attributes\ContentPath;
 use JacobJoergensen\LaravelPaper\Attributes\Disk;
 use JacobJoergensen\LaravelPaper\Attributes\Driver;
-use JacobJoergensen\LaravelPaper\Contracts\PaperModel;
-use JacobJoergensen\LaravelPaper\Paper;
+use JacobJoergensen\LaravelPaper\PaperModel;
 
 #[Driver('markdown')]
 #[ContentPath('articles')]
 #[Disk('paper')]
-final class Article extends Model implements PaperModel
+final class Article extends PaperModel
 {
-    use Paper;
-
     /** @var list<string> */
     protected $guarded = [];
 }