+++ tests/Fixtures/BrokenModel.php
@@ -4,18 +4,14 @@ declare(strict_types=1);
 
 namespace JacobJoergensen\LaravelPaper\Tests\Fixtures;
 
-use Illuminate\Database\Eloquent\Model;
 use JacobJoergensen\LaravelPaper\Attributes\ContentPath;
 use JacobJoergensen\LaravelPaper\Attributes\Driver;
-use JacobJoergensen\LaravelPaper\Contracts\PaperModel;
-use JacobJoergensen\LaravelPaper\Paper;
+use JacobJoergensen\LaravelPaper\PaperModel;
 
 #[Driver('markdown')]
 #[ContentPath('tests/content/invalid')]
-final class BrokenModel extends Model implements PaperModel
+final class BrokenModel extends PaperModel
 {
-    use Paper;
-
     /** @var list<string> */
     protected $guarded = [];